requirements.txt
configspace
numpy
python >= 3.7
pytorch == 1.9
pillow
simplejson
tensorboard
yacs